Rename AwsExternalResource to HadoopConfigurationLocalStackUtil
